moved submodule functionality into own files
